make sample operators in constraint::operators instead of in build_spec
